+++ styles/app/RootLayout.js
@@ -1,10 +1,10 @@
-import './RootLayout.css'
+import './RootLayout.css';
 
-var body = 'RootLayout_body__1myxvav0'
-var container = 'RootLayout_container__1myxvav4'
-var footer = 'RootLayout_footer__1myxvav3'
-var header = 'RootLayout_header__1myxvav2'
-var main = 'RootLayout_main__1myxvav1'
-var section = 'RootLayout_section__1myxvav5'
+var body = 'RootLayout_body__1myxvav0';
+var container = 'RootLayout_container__1myxvav4';
+var footer = 'RootLayout_footer__1myxvav3';
+var header = 'RootLayout_header__1myxvav2';
+var main = 'RootLayout_main__1myxvav1';
+var section = 'RootLayout_section__1myxvav5';
 
-export { body, container, footer, header, main, section }
+export { body, container, footer, header, main, section };
