+++ styles/app/aside/colors/colors.js
@@ -1,6 +1,6 @@
-import './colors.css'
+import './colors.css';
 
-var swatch = 'colors_swatch__1zpl8u1'
-var wrap = 'colors_wrap__1zpl8u0'
+var swatch = 'colors_swatch__1zpl8u1';
+var wrap = 'colors_wrap__1zpl8u0';
 
-export { swatch, wrap }
+export { swatch, wrap };
