+++ styles/app/random/colors/colors.css
@@ -1,20 +1,20 @@
-.colors_wrap__1zpl8u0 {
+.colors_wrap__1nkq4tu0 {
   display: grid;
   grid-template-columns: 1fr;
   grid-gap: var(--space-4__1ngcs9h5);
 }
-.colors_swatch__1zpl8u1 {
+.colors_swatch__1nkq4tu1 {
   background-color: var(--bg-color);
   border-radius: var(--radii-small__1ngcs9ht);
   padding: var(--space-4__1ngcs9h5);
 }
 @media screen and (min-width: 40em) {
-  .colors_wrap__1zpl8u0 {
+  .colors_wrap__1nkq4tu0 {
     grid-template-columns: 1fr 1fr;
   }
 }
 @media screen and (min-width: 60em) {
-  .colors_wrap__1zpl8u0 {
+  .colors_wrap__1nkq4tu0 {
     grid-template-columns: 1fr 1fr 1fr;
   }
 }
