+++ styles/app/random/colors/colors.js
@@ -1,0 +1,6 @@
+import './colors.css';
+
+var swatch = 'colors_swatch__1nkq4tu1';
+var wrap = 'colors_wrap__1nkq4tu0';
+
+export { swatch, wrap };
