+++ styles/ui/CodeBlock.js
@@ -1,5 +1,5 @@
-import './CodeBlock.css'
+import './CodeBlock.css';
 
-var code = 'CodeBlock_code__tzgmlm0'
+var code = 'CodeBlock_code__tzgmlm0';
 
-export { code }
+export { code };
