+++ styles/ui/Footnotes.js
@@ -1,5 +1,5 @@
-import './Footnotes.css'
+import './Footnotes.css';
 
-var footnote = 'Footnotes_footnote__x4qpcp0'
+var footnote = 'Footnotes_footnote__x4qpcp0';
 
-export { footnote }
+export { footnote };
